Revert "Process Square payments using webhooks (#1572)"
This reverts commit 7180020ee5dc5a5805a2ff216869593c736ff130.
