+++ .env
@@ -13,13 +13,6 @@ SQUARE_ACCESS_TOKEN=SQACCESSTOKEN1234
 SQUARE_LOCATION_ID=SQLOCID1234
 SQUARE_ENVIRONMENT=sandbox
 
-# Settings for the order webhook (to be notified when an order is created)
-SQUARE_PAYMENT_WEBHOOK_NOTIFICATION_URL = "https://example.com/square/callback"
-SQUARE_PAYMENT_WEBHOOK_SIGNATURE_KEY = "SQ_signature"
-
-# Delay redirects to Square to account for slowness in the sandbox environment
-SQUARE_SLOW_SANDBOX_DELAY=off
-
 ## Google Calendar (see README for details)
 # 
 # Path to Google service account credentials (used in prod)
