+++ .env.test
@@ -2,6 +2,3 @@ HEADLESS=true
 VOLUNTEER_SLOT_CALENDAR_ID=volcalid
 APPOINTMENT_SLOT_CALENDAR_ID=appcalid
 FEATURE_NEW_APPOINTMENTS_PAGE=false
-
-SQUARE_PAYMENT_WEBHOOK_NOTIFICATION_URL = "https://example.com/square/callback"
-SQUARE_PAYMENT_WEBHOOK_SIGNATURE_KEY = "SQ_signature"
