+++ app/controllers/renewal/confirmations_controller.rb
@@ -2,7 +2,7 @@ module Renewal
   class ConfirmationsController < BaseController
     def show
       activate_step(:complete)
-      @payment = session[:payment]
+      @amount = Money.new(session[:amount]) if session.key?(:amount)
     end
   end
 end
