+++ app/controllers/signup/payments_controller.rb
@@ -18,14 +18,10 @@ module Signup
 
       result = checkout.checkout_url(amount: @form.amount, email: @member.email, return_to: callback_signup_payments_url, member_id: @member.id, date: Date.current)
       if result.success?
-        if checkout.slow_sandbox_environment?
-          # It takes the sandbox a while to be ready
-          sleep 2
-        end
         redirect_to result.value, status: :see_other, allow_other_host: true
       else
-        Appsignal.set_error(result.error)
-        Rails.logger.error(result.error)
+        errors = result.error
+        Rails.logger.error(errors)
         flash[:error] = "There was a problem connecting to our payment processor."
         redirect_to new_signup_payment_url, status: :see_other
       end
@@ -39,8 +35,36 @@ module Signup
     end
 
     def callback
+      order_id = params[:orderId]
+      result = checkout.fetch_order(order_id: order_id)
+      session[:attempts] ||= 0
+
+      if result.success?
+        amount = result.value
+        Membership.create_for_member(@member, amount: amount, square_transaction_id: order_id, source: "square")
+        MemberMailer.with(member: @member, amount: amount.cents).welcome_message.deliver_later
+
+        reset_session
+        session[:amount] = amount.cents
+
+        redirect_to signup_confirmation_url, status: :see_other
+        return
+      end
+
+      errors = result.error
+
+      if errors.first[:code] == "NOT_FOUND"
+        # Give Square a little while for the transaction data to be available
+        session[:attempts] += 1
+        if session[:attempts] <= 10
+          render :wait, layout: nil
+          return
+        end
+      end
+
+      Rails.logger.error(errors)
       reset_session
-      session[:payment] = true
+      flash[:error] = "There was an error processing your payment. Please come into the library to complete signup."
       redirect_to signup_confirmation_url, status: :see_other
     end
 
@@ -51,7 +75,11 @@ module Signup
     end
 
     def checkout
-      SquareCheckout::Client.from_env
+      SquareCheckout.new(
+        access_token: ENV.fetch("SQUARE_ACCESS_TOKEN"),
+        location_id: ENV.fetch("SQUARE_LOCATION_ID"),
+        environment: ENV.fetch("SQUARE_ENVIRONMENT")
+      )
     end
 
     def are_payments_enabled?
