--- app/lib/square_checkout/square_error.rb
@@ -1,3 +1,0 @@
-module SquareCheckout
-  class SquareError < StandardError; end
-end
