+++ app/views/admin/members/memberships/index.html.erb
@@ -44,11 +44,7 @@
           <td><%= adjustment.amount.format %></td>
           <td>
             <% if adjustment.payment? %>
-              <% if adjustment.square_transaction_id %>
-                <%= link_to "Square payment", square_order_url(adjustment.square_transaction_id), target: "_blank" %>
-              <% else %>
-                <%= adjustment.payment_source.capitalize %> payment
-              <% end %>
+              <%= adjustment.payment_source.capitalize %> payment
             <% else %>
               <%= adjustment.kind.capitalize %>
             <% end %>
