+++ app/views/renewal/confirmations/show.html.erb
@@ -7,8 +7,8 @@
     <div class="verification-instructions">
       <p>Thank you for renewing your Chicago Tool Library membership.</p>
 
-      <% if @payment %>
-        <p><strong>Your payment is being processed and applied to your account.</strong></p>
+      <% if @amount %>
+        <p><strong>Your payment of <%= @amount.format %> has been processed and applied to your account.</strong></p>
       <% end %>
     </div>
     <p>If you have any questions, please email us at <a href="mailto:team@chicagotoollibrary.org">team@chicagotoollibrary.org</a>.</p>
@@ -18,6 +18,6 @@
 
 <div class="columns">
   <div class="column col-8 col-sm-12 col-mx-auto mt-2">
-    <%= link_to "Back to your account", account_home_url, class: "btn btn-block btn-primary" %>
+    <%= link_to "Back to your account", account_home_url, class: "btn btn-block" %>
   </div>
 </div>
