+++ app/views/signup/confirmations/show.html.erb
@@ -16,6 +16,6 @@
 
 <div class="columns">
   <div class="column col-8 col-sm-12 col-mx-auto mt-2">
-    <%= link_to "Back to Signup Page", signup_url, class: "btn btn-block btn-primary" %>
+    <%= link_to "Back to Signup Page", signup_url, class: "btn btn-block" %>
   </div>
 </div>
