+++ config/routes.rb
@@ -224,7 +224,6 @@ Rails.application.routes.draw do
   resources :homepage, only: [:index, :create]
 
   post "/twilio/callback", to: "twilio#callback"
-  post "/square/callback", to: "square#callback"
 
   # Mount dashboards for admins
   authenticate :user, ->(user) { user.admin? } do
